+++ src/FerriteViz.jl
@@ -12,6 +12,7 @@ abstract type AbstractPlotter end
 abstract type AbstractFilter end
 
 include("tessellation.jl")
+include("qptessellation.jl")
 include("dataset.jl")
 include("gradient.jl")
 include("filters.jl")
@@ -20,7 +21,7 @@ include("viewer.jl")
 
 export FEData, apply
 export set_point_data!, set_cell_data!
-export WarpByVector, Gradient, CrinkleClip, Refine, FirstOrderRefinement
+export WarpByVector, Gradient, CrinkleClip, Refine, FirstOrderRefinement, QuadraturePointData
 export Component, Magnitude, Norm1, VonMises, Deviator, Threshold, Derive
 export ClipPlane, vonmises
 export solutionplot, solutionplot!, cellplot, cellplot!, meshplot, meshplot!,